+++ app/src/main/java/com/app/wi_fi_direct/adapters/PeersAdapter.java
@@ -1,4 +1,4 @@
-package com.app.wi_fi_direct;
+package com.app.wi_fi_direct.adapters;
 
 import android.app.Activity;
 import android.content.Context;
@@ -11,6 +11,7 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Toast;
 
+import com.app.wi_fi_direct.R;
 import com.app.wi_fi_direct.helpers.ConnectPeer;
 
 import java.util.List;
