+++ app/src/main/java/com/app/wi_fi_direct/helpers/MyBroadcastReciever.java
@@ -1,4 +1,4 @@
-package com.app.wi_fi_direct;
+package com.app.wi_fi_direct.helpers;
 
 import android.content.BroadcastReceiver;
 import android.content.Context;
