--- app/src/main/java/com/app/wi_fi_direct/models/DeviceModel.java
@@ -1,17 +1,0 @@
-package com.app.wi_fi_direct.models;
-
-public class DeviceModel {
-  private String title;
-
-  public DeviceModel(String title) {
-    this.title = title;
-  }
-
-  public String getTitle() {
-    return title;
-  }
-
-  public void setTitle(String title) {
-    this.title = title;
-  }
-}
