--- app/src/main/res/layout/activity_recieve.xml
@@ -1,96 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_height="match_parent"
-    android:layout_width="match_parent"
-    android:gravity="center_horizontal"
-    android:orientation="vertical">
-
-    <TextView
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:text="@string/send_activity_first_heading"
-        android:gravity="center_horizontal"
-        android:textSize="15dp"
-        android:background="@color/textBackColor"
-        android:textColor="@color/textColor"
-        />
-
-    <View
-        android:layout_width="match_parent"
-        android:layout_height="15dp"/>
-
-    <TextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:id="@+id/recieve_name"
-        android:text="My Device"
-        android:textSize="20dp"
-        android:textColor="@color/textColor"
-         />
-
-    <View
-        android:layout_width="match_parent"
-        android:layout_height="15dp"/>
-
-    <TextView
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:text="@string/receive_activity_second_heading"
-        android:gravity="center_horizontal"
-        android:textSize="15dp"
-        android:background="@color/textBackColor"
-        android:textColor="@color/textColor"
-        />
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:orientation="vertical"
-        android:gravity="left"
-        android:id="@+id/currentFile"
-        android:clickable="true"
-        android:background="?attr/selectableItemBackground">
-
-        <View
-            android:layout_width="match_parent"
-            android:layout_height="15dp"/>
-
-    <TextView
-        android:layout_width="match_parent"
-        android:layout_height="30dp"
-        android:id="@+id/fileName"
-        android:textSize="20dp"
-        android:text=""
-        android:textColor="@color/textColor"
-        />
-    <TextView
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:id="@+id/dateAndTime"
-        android:textSize="15dp"/>
-
-        <ProgressBar
-            android:layout_width="match_parent"
-            android:id="@+id/receiveProgressBar"
-            android:layout_height="5dp"
-
-            style="@style/Widget.AppCompat.ProgressBar.Horizontal"/>
-
-    </LinearLayout>
-
-    <TextView
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:text="@string/receive_activity_third_heading"
-        android:gravity="center_horizontal"
-        android:textSize="15dp"
-        android:background="@color/textBackColor"
-        android:textColor="@color/textColor"
-        />
-
-    <android.support.v7.widget.RecyclerView
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:id="@+id/receivedRecycler"/>
-
-</LinearLayout>
