--- app/src/main/res/layout/file_list_item.xml
@@ -1,57 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:orientation="vertical" android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:gravity="left"
-    android:clickable="true"
-    android:id="@+id/receivedFile"
-    android:background="?attr/selectableItemBackground">
-
-    <LinearLayout
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:orientation="horizontal">
-
-        <View
-            android:layout_width="10dp"
-            android:layout_height="match_parent"/>
-
-    <LinearLayout
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:orientation="vertical">
-
-    <View
-        android:layout_width="match_parent"
-        android:layout_height="10dp"
-        />
-
-    <TextView
-        android:layout_width="match_parent"
-        android:layout_height="30dp"
-        android:textSize="20dp"
-        android:id="@+id/receivedFileName"
-        android:textColor="@color/textColor"/>
-
-    <TextView
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:id="@+id/receivedFileDate"/>
-
-    <View
-        android:layout_width="match_parent"
-        android:layout_height="10dp"
-        />
-
-        </LinearLayout>
-
-        </LinearLayout>
-
-    <View
-        android:layout_width="match_parent"
-        android:layout_height="1dp"
-        android:background="@color/textBackColor"
-        />
-
-
-</LinearLayout>
