--- app/src/main/res/layout/peers_list_item.xml
@@ -1,39 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:orientation="vertical" android:layout_width="match_parent"
-    android:layout_height="60dp"
-    android:id="@+id/peerListItem"
-    android:gravity="center_horizontal">
-
-    <View
-        android:layout_width="match_parent"
-        android:layout_height="10dp"/>
-
-    <LinearLayout
-        android:orientation="vertical"
-        android:clickable="true"
-        android:layout_width="200dp"
-        android:layout_height="50dp"
-        android:id="@+id/peerButton"
-        android:background="?attr/selectableItemBackground">
-
-        <TextView
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:id="@+id/peerName"
-            android:gravity="center_horizontal"
-            android:textSize="20dp"/>
-
-        <View
-            android:layout_width="match_parent"
-            android:layout_height="5dp"/>
-
-        <TextView
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:id="@+id/peerAddress"
-            android:gravity="center_horizontal"/>
-
-    </LinearLayout>
-
-</LinearLayout>
